Rename: 파일구조 및 파일명 수정
컴포넌트 파일구조 갈아엎음
