+++ myStarbucks/components/widgets/SearchResultCategory.tsx
