+++ myStarbucks/components/widgets/SearchResultInfo.tsx
